+++ tests/test_octodns_manager.py
@@ -8,6 +8,7 @@ from __future__ import absolute_import, division, print_function, \
 from os import environ
 from os.path import dirname, join
 
+from octodns import __VERSION__
 from octodns.manager import _AggregateTarget, MainThreadExecutor, Manager, \
     ManagerException
 from octodns.processor.base import BaseProcessor
@@ -425,7 +426,7 @@ class TestManager(TestCase):
         plan_output_mock = MagicMock()
         plan_output_class_mock = MagicMock()
         plan_output_class_mock.return_value = plan_output_mock
-        mock.return_value = plan_output_class_mock
+        mock.return_value = (plan_output_class_mock, 'ignored', 'ignored')
         fh_mock = MagicMock()
 
         Manager(get_config_filename('plan-output-filehandle.yaml')
@@ -441,7 +442,7 @@ class TestManager(TestCase):
     def test_processor_config(self):
         # Smoke test loading a valid config
         manager = Manager(get_config_filename('processors.yaml'))
-        self.assertEqual(['noop'], list(manager.processors.keys()))
+        self.assertEqual(['noop', 'test'], list(manager.processors.keys()))
         # This zone specifies a valid processor
         manager.sync(['unit.tests.'])
 
@@ -523,6 +524,29 @@ class TestManager(TestCase):
         # We planned a delete again, but this time removed it from the plan, so
         # no plans
         self.assertFalse(plans)
+
+    def test_try_version(self):
+        manager = Manager(get_config_filename('simple.yaml'))
+
+        class DummyModule(object):
+            __VERSION__ = '2.3.4'
+
+        dummy_module = DummyModule()
+
+        # use importlib.metadata.version
+        self.assertTrue(__VERSION__,
+                        manager._try_version('octodns',
+                                             module=dummy_module,
+                                             version='1.2.3'))
+
+        # use module
+        self.assertTrue(manager._try_version('doesnt-exist',
+                                             module=dummy_module))
+
+        # fall back to version, preferred over module
+        self.assertEqual('1.2.3', manager._try_version('doesnt-exist',
+                                                       module=dummy_module,
+                                                       version='1.2.3', ))
 
 
 class TestMainThreadExecutor(TestCase):